moved endpoint.ts to `./common` folder
